% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{avgNcut}
\alias{avgNcut}
\title{Average Normalized N-Cut}
\usage{
avgNcut(g, clusters, verb = FALSE, ignore.zero = FALSE)
}
\arguments{
\item{g}{graph object of class \code{\link{ELgraph}}, the base graph.}

\item{clusters}{vector of size \code{n}. Element \code{i} indicates the cluster of node \code{v_i}.}

\item{verb}{boolean value: if \code{TRUE}, the function prints the n-cut of all clusters.}

\item{ignore.zero}{boolean value: if \code{TRUE}, clusters with internal degree 0 are ignored.}
}
\value{
integer: the average normalized n-cut of a clustering.
}
\description{
This function computes the average normalized n-cut of a clustering.
}
\examples{
g = generateEdgeList(n=1000)
res = ML_RMCL(g)
avgNcut(g, res, verb = TRUE)
}
\keyword{average}
\keyword{n-cut}
\keyword{ncut}
\keyword{normalized}
