% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphRead.R
\name{createGraph}
\alias{createGraph}
\title{Graph Creation From File}
\usage{
createGraph(path, nskip = 0, directed = TRUE, norm = TRUE,
  inverse.direction = TRUE)
}
\arguments{
\item{path}{string: the path to the file containing the edge list.}

\item{nskip}{integer: number of lines to skip before starting to read the edges.}

\item{directed}{boolean: value describing if the edges are directed.}

\item{norm}{boolean: if \code{TRUE}, translate the vertices names so that they are from 1 to n.}

\item{inverse.direction}{boolean: if \code{FALSE}, labels the first column \code{i} and the second \code{j}, otherwise the inverse.
this option is used when creating a graph for the ML-RMCL function, because it understand the edges going from \code{j} to \code{i}.}
}
\value{
A graph object of class \code{\link{ELgraph}} from the edge list found at \code{path}
}
\description{
This function creates a graph object from a file containing the edges
}
\examples{
g = createGraph("graphs/myEdges.txt")
}
\keyword{edge}
\keyword{file}
\keyword{graph}
\keyword{initialization}
\keyword{list}
