% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphClass.R
\name{ELgraph}
\alias{ELgraph}
\title{A Graph Object}
\usage{
ELgraph(edgeList, nodeList = NULL, nodeMap = NULL, clusters = NULL,
  directed = TRUE)
}
\arguments{
\item{edgeList}{\code{data.table} with a column i and a cloumn j of edges endpoints}

\item{nodeList}{vector of node names. if \code{NULL}, it is derived from the edge list}

\item{nodeMap}{table of nodes mapped. Can be used when the graph is the result of coarsening}

\item{clusters}{vector of size \code{length(nodeList)} of truth cluster belonging.}

\item{directed}{boolean value describing if the edges are directed.}
}
\value{
A graph object
}
\description{
This function allows you to create graph objects from an edge list.
}
\examples{
g = ELgraph(data.table("i"=c(2,5,6,3,2), "j" =c(4,4,3,1,5)))
}
\keyword{graphs}
\keyword{object}
