% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateEdgeList}
\alias{generateEdgeList}
\title{Synthetic Graph Generation}
\usage{
generateEdgeList(k = 2, aid = 100, aed = 10, n = 1000, prob = NULL,
  directed = FALSE)
}
\arguments{
\item{k}{integer: the number of clusters.}

\item{aid}{integer: average intra-cluster degree. Poisson parameters for number of neighbors within the cluster.}

\item{aed}{integer: average inter-cluster degree. Poisson parameters for number of neighbors outside the cluster.}

\item{n}{integer: number of vertices in the graph.}

\item{prob}{vector of size \code{k} with probabilities that a node belongs to a cluster. If omitted, all probabilities are equal \code{1/k}.}

\item{directed}{boolean value describing if the edges are directed.}
}
\value{
A graph object of class \code{\link{ELgraph}} with truth clusters.
}
\description{
This function creates a synthetic graph with truth clusters properties.
}
\examples{
g = generateEdgeList(n=1000)
}
\keyword{generation}
\keyword{graph}
