% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphClass.R
\name{getSparseMat}
\alias{getSparseMat}
\title{Extract the Adjancency Sparse Matrix}
\usage{
getSparseMat(graph)
}
\arguments{
\item{graph}{The graph from which to obtain the matrix}
}
\value{
A matrix of sparse format from package Matrix with entries corresponding to the edge list of the given graph
}
\description{
This function allows you to get the adjacency matrix from a graph object.
}
\examples{
M = getSparseMat(g)
}
\keyword{adjacency}
\keyword{matrix}
\keyword{sparse}
